pyproject.toml
[project]
name = "ai-crypto-trading-system"
version = "0.1.0"
description = ""
authors = [
    {name = "google-labs-jules[bot]",email = "161369871+google-labs-jules[bot]@users.noreply.github.com"}
]
readme = "README.md"
requires-python = "~3.10"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (~=1.26.0)", # Pinning to a known stable series for pandas 2.2
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "alembic (>=1.16.1,<2.0.0)"
]


[tool.poetry]
packages = [
    { include = "trading_system" }
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
